Added the beginnings of Shared Object Relocation code. Added RelocationType enumeration with Intel 386 values. Added ELFSharedObject.BaseAddress field. Refactored for spelling correction: VirtMemManager.FindFreeVirtsPage to FindFreeVirtPage.
Former-commit-id: bb6d533ea7906b47277f82b6a4408654d665c8ab

+++ Kernel/Libraries/Kernel.Core/Processes/ELF/ELFFile.cs
@@ -208,7 +208,7 @@ namespace Kernel.Core.Processes.ELF
 
                             Console.Default.WriteLine("     - Relocation : ");
                             Console.Default.Write("         - Type : ");
-                            Console.Default.WriteLine_AsDecimal(theRel.Type);
+                            Console.Default.WriteLine_AsDecimal((uint)theRel.Type);
                             Console.Default.Write("         - Symbol : ");
                             Console.Default.WriteLine_AsDecimal(theRel.Symbol);
                             Console.Default.Write("         - Offset : ");
@@ -237,7 +237,7 @@ namespace Kernel.Core.Processes.ELF
 
                             Console.Default.WriteLine("     - Relocation : ");
                             Console.Default.Write("         - Type : ");
-                            Console.Default.WriteLine_AsDecimal(theRel.Type);
+                            Console.Default.WriteLine_AsDecimal((uint)theRel.Type);
                             Console.Default.Write("         - Symbol : ");
                             Console.Default.WriteLine_AsDecimal(theRel.Symbol);
                             Console.Default.Write("         - Offset : ");
